refactor: separating unit and integration test
